feat(mpt): add feature-gated witness module (WitnessDb/WitnessInput)
Move WitnessDb, WitnessInput and the header-chain validation out of
stateless-executor/io.rs into openvm-mpt behind a new `witness` feature, so
downstream executors can reuse them instead of copying. Pure relocation; the
header-validation errors move to a dedicated `WitnessError`.
